+++ routes/generos.js
@@ -1,5 +1,6 @@
 const db = require('../modules/firebase');
 const express = require('express');
+const jwt = require('../utils/jwt');
 const router = express.Router();
 
 router.route('/generos').get(async function(req, res) {
@@ -15,28 +16,46 @@ router.route('/generos').get(async function(req, res) {
 });
 
 router.route('/generos').post(async function(req, res) {
-  const genero = {
-    nombre: req.body.nombre
-  };
-  const docRef = await db.collection('generos').add(genero);
+  if(jwt.validateToken){
+    const genero = {
+      nombre: req.body.nombre
+    };
+    const docRef = await db.collection('generos').add(genero);
 
-  res.json({ message: 'Genero creado', id: docRef.id });
+    res.json({ message: 'Genero creado', id: docRef.id });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/genero/:id').put(async function(req, res) {
-  const genero = {
-    nombre: req.body.nombre
-  }
-  
-  await db.collection('generos').doc(req.params.id).set(genero);
+  if(jwt.validateToken){
+    const genero = {
+      nombre: req.body.nombre
+    }
 
-  res.json({ message: 'Genero actualizado' });
+    await db.collection('generos').doc(req.params.id).set(genero);
+
+    res.json({ message: 'Genero actualizado' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 router.route('/genero/:id').delete(async function(req, res) {
-  await db.collection('generos').doc(req.params.id).delete();
-  
-  res.json({ message: 'Genero eliminado' });
+  if(jwt.validateToken){
+    await db.collection('generos').doc(req.params.id).delete();
+
+    res.json({ message: 'Genero eliminado' });
+  }
+  else{
+    res.json({"message": "no autorizado"});
+  }
+
 });
 
 module.exports = router;
